add update cancel and tips

+++ web/src/main/webapp/angularApp/controllers/userController.js
@@ -9,27 +9,44 @@ angular.module('myApp')
 
         $scope.deleteUser = function(event,id){
             var tr = $(event.target).parent().parent();
-            $http({
-                method:'DELETE',
-                url:'/web/angular/user/',
-                params:{'id':id}
-            }).success(function(){
-                tr.remove();
-            });
+
+            if(confirm("确定要删除该信息吗？")){
+                $http({
+                    method:'DELETE',
+                    url:'/web/angular/user/',
+                    params:{'id':id}
+                }).success(function(){
+                    tr.remove();
+                    alert("删除成功");
+                });
+            }
         };
 
         $scope.addUser=function(){
-            $http({
-                method:'POST',
-                url:'/web/angular/user/',
-                params:{
-                    'name':$scope.name,
-                    'password':$scope.password,
-                    'employee_id':$scope.employee
+
+            if($scope.name == null ||$scope.password == null || $scope.employee == null ){
+                alert("请输入完整注册信息");
+            }else{
+
+                $http({
+                    method:'POST',
+                    url:'/web/angular/user/',
+                    params:{
+                        'name':$scope.name,
+                        'password':$scope.password,
+                        'employee_id':$scope.employee
+                    }
+                }).success(function(result){
+                        if(result === "failed"){
+                            alert("该工号已占用，请输入正确工号");
+                        }else{
+                            $scope.users.push(result);
+                        }
                 }
-            }).success(function(user){
-                $scope.users.push(user);
-            });
+                ).error(function () {
+                    alert(" 操作有误，请正确输入");
+                });
+            }
         };
 
         $scope.update = [];
@@ -51,8 +68,20 @@ angular.module('myApp')
                     'employee_id':user.employee.id
                 }
             }).success(function(){
+            });
+        };
+
+        $scope.cancelUpdate = function($index){
+            $scope.update[$index] = true;
+            $http({
+                method:'GET',
+                url:'/web/angular/user/'
+            }).success(function(users){
+                $scope.users = users;
 
             });
+
+
         };
 
         //$scope.test= function () {
